Fix windows x64 build.
TBR=svenpanne@chromium.org

Review URL: https://codereview.chromium.org/280223002

git-svn-id: https://v8.googlecode.com/svn/branches/bleeding_edge@21263 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/d8.cc
@@ -1353,7 +1353,7 @@ void SourceGroup::WaitForThread() {
 
 
 void SetFlagsFromString(const char* flags) {
-  v8::V8::SetFlagsFromString(flags, strlen(flags));
+  v8::V8::SetFlagsFromString(flags, static_cast<int>(strlen(flags)));
 }
